// src/api/generated/models/publishersListParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { PublishersListSortBy } from './publishersListSortBy';
import type { PublishersListSortOrder } from './publishersListSortOrder';
import type { PublishersListUsageStatus } from './publishersListUsageStatus';

export type PublishersListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
search?: string;
usageStatus?: PublishersListUsageStatus;
createdFrom?: string;
createdTo?: string;
sortOrder?: PublishersListSortOrder;
sortBy?: PublishersListSortBy;
};